+++ CHANGE_LOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- 支持多头注意力；
+
 ## [0.0.4] - 2025-06-19
 
 ### Added